backend/pyproject.toml
[tool.poetry]
name = "frankfurtian-tens-thought-experiment"
version = "0.1.0"
description = ""
authors = ["Owen McAllister <owencmcallister@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.12.5"
fastapi = "^0.123.5"
uvicorn = "^0.38.0"
screeninfo = "^0.8.1"
pynput = "^1.8.1"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
